--- app/community/notice/error.tsx
@@ -1,17 +1,0 @@
-'use client';
-
-import { useEffect } from 'react';
-
-export default function Error({ error, reset }: { error: Error; reset: () => void }) {
-  useEffect(() => {
-    console.log(error.message);
-    console.log(error.stack);
-  }, [error]);
-
-  return (
-    <div>
-      <h2>Something went wrong!</h2>
-      <button onClick={() => reset()}>Try Again</button>
-    </div>
-  );
-}
